80% done
Menu method has been tweaked into working better but minor issue need to be fixed such as, the program going belly up as soon as a different data type is entered as a choice. Minor changes to outputting methods need to be done to make the program user friendly. File Handling components havent been touched yet but will work on it soon. There is a help method in mind to resolve common issues but will see where this leads.